[write standby] retry DB connections
At service start up time, there are chances that the networking service is being
restarted by interface-config service. When that happens, write_standby could
fail to make DB connections due to loopback interface is being reconfigured.

Since the reconfiguration is transient, add a few retries to the DB connections
so that the script won't fail unnecessarily

Signed-off-by: Ying Xie <ying.xie@microsoft.com>