Prefer dget to wget for some debian src downloads
- use dget instead of wget in some Makefiles
- switch related HTTP URL to HTTPS and official debian mirror when
  possible

TODO: migrate azure debian mirror to debian official ones (this might
require upgrading some debian packages version in rules/*.mk)
NB: A more complete solution with apt-get is proposed in
https://github.com/sonic-net/sonic-buildimage/pull/13381

SONiC images compilation is currently designed to be performed with a
non-root user under Ubuntu systems.
And the build process recompiles packages for various GNU+Linux Debian
distributions.
Though, there is no guarantee that debian repos are declared in
/etc/apt/sources.list on the Ubuntu host system used.
As a result, 'apt-get source' can not be directly used out-of-the-box.
Currently, the packages sources are retrieved from more or less official URLs
and locations with dget and wget.

As discussed in the following URL
https://serverfault.com/questions/447457/use-apt-get-source-on-a-debian-repo-without-using-etc-apt-source-list
dget is the simplest solution.

wget has several drawbacks from a maintenance and security standpoint:

- retrieving multiple URLs
- loose authentication of packages sources especially when retrieved with HTTP instead of HTTPS
  This is potentially vulnerable to MITM and can induce malicious code injection.

dget by default verifies package signature and can prevent most of such MITM attacks

Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>