Make changes to the transition mechanism
First, reply with the most recently received LACP PDU version. This is
so that if one side is sending both the old and new versions, then the
peer side will send both versions to keep both listeners happy. This
also allows for using the new version packets even when the retry count
is 3 on both sides. This can only be triggered remotely; there's no
local command to force a particular version to be used.

Second, reset the 60-second timer preventing the retry count from being
reset back to 3 every time a new version PDU is received. This keeps
that setting persistent for longer with the downside of a delayed
response to the retry count being reset to 3.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>