Added changes for chassis: 1. Allow traffic with source and destination as chassis eth1-midplane ip. Needed for Supervisor Redis-db connection after we load acl.json that has catch-all drop rule
2. Made multi_asic_ns_to_host_fwd as False for ACL service for External
Client. This flag is needed for service SSH and SNMP where traffic can
come in namespace over front-panel ports and we need to send the traffic
in host where corresponding docker/service are running. There is no
use-case of External client service for multi-asic as of now. Having
flag as True creates failure when we try to load acl.json.

Signed-off-by: Abhishek Dosi <abdosi@microsoft.com>