Removed extra 'ip addr del' on the lo inteface
The 'post-up ip addr del 127.0.0.1/8 dev lo' was thought to be required
when the loopback interface subnet mask was narrowed from /8 to /16.
After testing first reboot, subsequent reboots, config reloads, and
systemctl restart networking on linecards and supevisors (DNX VOQ
chassis), this cleanup command does not appear to be required. As such,
this change removes it as well as reverts
https://github.com/sonic-net/sonic-buildimage/pull/15080.