[Arista] Remove one pcie device accross platforms
These devices will not reliabily report the proper devid and vendorid
when reading it is read directly from the pci config space.
It can be read but shouldn't be compared against some fixed value like
the one stored in pcie.yaml.

Since this makes pcied unhappy, the simplest path forward is to just
remove this device from monitoring.