Fix building the SONiC slave container for QEMU-based build
The existing source of multiarch/debian-debootstrap doesn't appear to
have Bookworm-based images available. Since those images don't do
anything to the Debian container besides add QEMU to it (which we
overwrite anyways with the latest version of QEMU available from
multiarch/qemu-user-static, just take the platform-specific version of
the official Debian image and add QEMU to it.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>