[rsyslog]: Use RELP instead of UDP for forwarding from container to host
When the host's rsyslog is restarted (for example, to regenerate the
config after some changes, or as part of some automated script), there
is a chance that some syslog messages from the containers are lost. Most
of the time, this isn't an issue. However, if there are test cases that
expect all syslogs to be present (such as the advanced-reboot test
case), then this can cause a problem. Additionally, this could affect
debuggability of issues where a rsyslog restart happens in the middle.

There are two options for reliable message transport in rsyslog: TCP and
RELP. With TCP, while the protocol knows whether a syslog message has
been delivered or not, the application doesn't know, because there is no
feedback from the remote side saying the message was received. This
means that there is still a chance that messages could be lost when the
connection is broken (if, for example, the host rsyslog gets restarted),
because after the connection is established, the sender rsyslog (in the
container) doesn't know if the message has been received or not.

RELP instead adds a feedback mechanism where the remote side notifies
the sender whether the message has actually been received or not. This
makes it much less likely to lose a message. There is one known possible
case where a message (or messages) could be lost: the network is down,
and rsyslog gets restarted. This at least requires both the network and
rsyslog to have an issue, rather than just one. There is also a slim
possibility where a message could get duplicated; this should be mostly
fine (hopefully).

RELP does require that both sides are using a recent version of rsyslogd
(at least 7.3.16, which looks like it was released more than 10 years
ago), but since we use Debian on both the container and the host, it
should be fine.

Therefore, switch to using RELP when sending syslog messages from the
container to the host.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>