Install netaddr 0.10.1 in the Python 2 part of sonic-mgmt
Recently, netaddr 1.0.0 was released, which dropped Python 2 support
entirely. Because of this, an empty netaddr package is installed for
Python 2. Ansible still uses Python 2, so this needs to remain
functional. The Python 3 part of sonic-mgmt is fine.

Pin the version of netaddr installed in Python 2 to 0.10.1, the last
supported version.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>