Modify omrelp to store messages in memory if the server is unavailable
By default, just using omrelp doesn't hold log messages if the server
happens to be unavailable. This needs to be configured manually.

Configure an in-memory storage (of a linked list) that by default will
store up to 1000 messages (this appears to be a default value that can
be bumped up) if the server is unavailable. I'm assuming this will be
sufficient for most cases.

Assuming each message is 512 bytes (many of our messages will be smaller
than this), this will take up an additional 512kB of memory if 1000
messages are queues. If there are no messages queued, then no additional
space is taken up.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>