Add lcov to Bookworm slave container
The gcovr package used to have lcov as a runtime dependency. However,
that was removed in Debian bug 987818. However, autoconf's code coverage
macro file expects lcov and genhtml to be present.

It's possible that lcov and genhtml are actually not needed for
generating the report, and that gcovr is enough. For now, though, just
add lcov (which includes genhtml) into the slave container.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>