This commit adds:
A service 'fsio-sync' that calls a shell script 'fsio-sync.sh' that
parses the latest fsio reads/writes from /proc/diskstats and saves it
to disk using 'fsio-rw-sync' script, as well as updates the StateDB
upon successful parse and sync.

Changes that copy these files to the appropriate locations in the image.

Changes that instruct the build_template to copy 'fsio-rw-sync' into
the appr. location in the pmon container