Use apt preferences to prefer packages from the mirror over the main repos
Currently, if snapshots are enabled, then packages from the mirror and
packages from the main repos have the same priority. This becomes
problematic if the main repos have a package that's of a newer version
than the snapshots, and something wants to install that newer version,
even though it could be installed with that older version. For example,
if a -dev package is being installed, and there's a newer version of the
-dev and the main package in the main repos (compared to the snapshots),
then apt will try to install the newer -dev package, which will have a
dependency on the newer main package, but the older main package will
still be installed. It could instead try to install the older -dev
package from the snapshots, which would be installable.

To fix this, use apt preferences to give the pacakges in the snapshot
repo a higher priority than the ones in the main repo. This should be
enough to convince apt to use packages from the snapshot repo whenever
possible, but it still allows packages from the main repo to be
installed, in case some change comes in that requires that for
whatever reason.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>