(Properly) upgrade docker-restapi to Bookworm, and add a debug container
Upgrade docker-restapi container to Bookworm. It's worth noting that
because this container was not in the SONIC_BULLSEYE_DOCKERS list, the
binaries/libraries that got packaged here came from the Bookworm build,
but the base layer was for Bullseye, so it was already in some hybrid
state.

Also, include the autogenerated dbgsym package for restapi in
`rules/restapi.mk`, and add a debug container for restapi.

Also, enable caching of docker-restapi (not sure why it wasn't enabled
before).

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>