Update save disable code for new Redis version
With the upgrade of docker-database to Bookworm and the new Redis
version that was brought in, database saves are enabled by default.
However, we don't use (and maybe don't want, at this time) those
database saves.

Restore the old behavior by modifying `/etc/redis/redis.conf` to not
save at all.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>