[FRR] Fix syntax error for get correctly value
What I did :
Fix the syntax error for get correctly value

Why I did :
There is a syntax error, so the value cannot be retrieved correctly.
As a result, a wrong value (null) is written to the specified Python file.

This error causes the specified Python file to not know which Python version
to use for execution.