Enable iburst option for NTP servers loaded from minigraph
In #15058, the NTP server configuration was modified to add additional
options, such as conditionally enabling iburst, specifying the
association type, and specifying the NTP version. One side effect of
this was that the iburst option which was previously always enabled now
requires it to be explicitly enabled in the config_db.

To restore the old behavior, when loading from minigraph, add
`"iburst": "true"` for each NTP server loaded from minigraph.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>