[Mellanox] implement state machine for always firmware control ports (#19473)
- Why I did it
When module host management is enabled, all modules are managed by the same state machine which initializes all modules/ports. However, on 5600 and 5400, there is a type of port which we called "service port". Those ports are always under firmware control. In that case, those ports should be managed by a different state machine. This PR is to implement the state machine for always firmware control ports.

- How I did it
Implement a state machine for always firmware control ports.

- How to verify it
Manual test
unit test