Pin pip to version 24.2, and update disable-non-manylinux.patch
Pip has been recently updated to 24.2. For armhf, there is a patch file
that gets applied to disable the use of manylinux wheels, to avoid the
use of packages that might be using a newer glibc version than what is
available on the system. This patch file was designed for 24.0, but
can't be applied to 24.2.

Update the patch file to apply on 24.2, and make sure version 24.2 of
pip gets installed.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>