Add an artificial dependency between two modules for centec-arm64
For centec-arm64, both platform-modules-fs and platform-modules-e530 try
to compile the same directory of kernel sources. If both compilations
are running at the same time, then this is going to lead to errors.

Fix this by introducing an artificial dependency between the two so that
the compilations are forced to be serialized.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>