Change whl and deb package installations to happen within chroot
Currently, for deb package installations, the --root argument is passed
into dpkg to install packages in a custom root directory. However, this
doesn't seem to do a full chroot, because there are some user/group name
checks that appear to be done before the chroot has taken place. This
becomes an issue with the chrony installation; if chrony doesn't happen
to be installed on the host system, then installing any pacakge after
chrony is installed in the fsroot that we're building will fail with the
following error:

```
+ sudo DEBIAN_FRONTEND=noninteractive dpkg --root=./fsroot-vs -i target/debs/bookworm/kdump-tools_1.8.1_amd64.deb
dpkg: unrecoverable fatal error, aborting:
 unknown system group '_chrony' in statoverride file; the system group got removed
before the override, which is most probably a packaging bug, to recover you
can remove the override manually with dpkg-statoverride
```

Therefore, for deb (and whl) package installations, define a bash
function that will copy the file into the fsroot, chroot into it,
install the pacakge, and then remove it. Replace all existing deb and
whl installations with code to call this function, so that it's
consolidated.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>