Update docker to 27.2.1 (26.1.4 for Buster), and enable containerd storage
With the new approach of building the images (where the entire final
rootfs is copied into the second stage), if the system building the
containers is using the overlay2 storage driver (which is the default)
and is able to use native diffs (which could be true if
CONFIG_OVERLAY_FS_REDIRECT_DIR isn't enabled in the kernel), then the
final result of the image will be different than if naive diffs (where
Docker compares the metadata of each file and, if needed, the contents
to find out if something has changed) were used. Specifically, with
native diffs, each container would be much larger, since technically
speaking, the whole rootfs is being written to, even if the content ends
up the same. This appears to be a known issue (in some form), and
workarounds are being thought of in moby/moby#35280.

As a workaround for SONiC, for the docker daemon that we start up in the
slave container, enable the containerd image store. The overlayfs
storage driver in containerd always takes naive diffs, and never native
diffs (the code for this isn't there at all). This makes sure that the
diffing method is the same regardless of host system.

The caveat here is that if someone builds SONiC with the native Docker
daemon (instead of one that we start in the slave container), then their
image build would be significantly larger. For this, I can't think of a
fix.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>