Set the nofile ulimit for the slave container to 1024:1048576
It seems that on the Bullseye slave container (not sure about Buster),
the nofile ulimit is set to 1048576:1048576 (as in, 1048576 for both the
soft and hard limit). However, the Docker startup script in version 25
and newer sets the hard limit to 524288 (because of
moby/moby#c8930105b), which fails because then the soft limit will be
higher than the hard limit, which doesn't make sense.  However, on a
Bookworm slave container, the nofile ulimit is set to
1024:1048576, and the startup script's ulimit command goes through.

A simple workaround would be to explicitly set the nofile ulimit to be
1024:1048576 for all slave containers. However, sonic-swss's tests needs
more than 1024 file descriptors open, because the test code doesn't
clean up file descriptors at the end of each test case/test suite. This
results in FD leaks.

Therefore, set the ulimit to 524288:1048576, so that Docker's startup
script can lower it to 524288 and swss can open file descriptors.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>