The hostname-config.sh script rewrites the hostname and appends the 127.0.0.1 new hostname mapping to the /etc/hosts list. However, it does not handle the case where the file does not exist, resulting in the hostname not being updated in /etc/hosts.
When /etc/hosts is missing, it causes the sudo command to be blocked
for a while and displays an "unable to resolve host" error message.

This patch adds a workaround in hostname-config.sh to create a
default /etc/hosts if it does not exist, ensuring the hostname is
updated properly.

Signed-off-by: gord_chen <gord_chen@edge-core.com>